+++ src/handler-dispatcher.ts
@@ -8,7 +8,7 @@ import { PRCommentHandler } from "./handlers/pr-comment";
 import type { Logger } from "./infra/logger";
 import type { ActionOutputs, HandlerConfig } from "./config/handler-config";
 import type { AppConfig } from "./config/app-config";
-import type { RouteResult } from "./webhook-router";
+import type { RouteResult } from "./webhooks/github/router";
 
 // ── Public interface ──────────────────────────────────────────────────────────
