+++ src/main.ts
@@ -3,7 +3,7 @@ import { Octokit } from "@octokit/rest";
 import { loadConfig } from "./config/app-config";
 import { createLogger } from "./infra/logger";
 import { RealCoderClient } from "./services/coder/client";
-import { WebhookRouter } from "./webhook-router";
+import { WebhookRouter } from "./webhooks/github/router";
 import { HandlerDispatcher } from "./handler-dispatcher";
 import { createApp } from "./http/server";
